+++ src/lib/config/es5.js
@@ -1,3 +1,0 @@
-export default {
-    extends: 'plugin:itgalaxy/core'
-};